+++ test/libsolidity/syntaxTests/empty_struct.sol
@@ -2,4 +2,4 @@ contract test {
 	struct A {}
 }
 // ----
-// SyntaxError: Defining empty structs is disallowed.
+// Warning: Defining empty structs is deprecated.