+++ test/libsolidity/syntaxTests/empty_struct_050.sol
@@ -1,0 +1,6 @@
+pragma experimental "v0.5.0";
+contract test {
+	struct A {}
+}
+// ----
+// SyntaxError: Defining empty structs is disallowed.